refactor: move certificate log into the pods directory
