+++ apps/drec-api/src/drec.module.spec.ts
@@ -21,7 +21,7 @@ import { CheckCertificateIssueDateLogForDeviceGroupEntity } from './pods/device-
 import { SDGBenefit } from './pods/sdgbenefit/sdgbenefit.entity';
 import { UserLoginSessionEntity } from './pods/user/user_login_session.entity';
 import { DeviceLateOngoingIssueCertificateEntity } from './pods/device/device_lateongoing_certificate.entity';
-import { CertificateLogModule } from './lib/certificate-log/certificate-log.module';
+import { CertificateLogModule } from './pods/certificate-log/certificate-log.module';
 import { SDGBenefitModule } from './pods/sdgbenefit/sdgbenefit.module';
 import { CountryCodeModule } from './pods/countrycode/countrycode.module';
 import { PermissionModule } from './pods/permission/permission.module';
